requirements.txt
Nikola
webassets
jinja2
markdown
versa
amara3-iri
amara3-xml

